backend/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
requests==2.26.0
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.2
Flask-Talisman==0.8.1
werkzeug==2.0.3
SQLAlchemy==1.4.47
psycopg2-binary==2.9.3 